requirements.txt
streamlit
langchain
openai
pdfplumber
python-docx
langchain_community